Better fix for symlinks
Better fix for symlinks

This is a better fix for finding symlinks.  This uses the `find` option "-L" to follow the symlink, but return the path using the symlink.  Although, it may be better to use the "-H" option to return the path to the linked file as it will keep `list_path` from returning the two identical files if the symlink points to a file in the same search path.