Version 19.3.0 (#15)
* Updates for POSIX sed

This update allows systems that use a POSIX version of sed (Max OS X).

* Cleanup whitespace/tabs in list_paths

* Remove all ^./ from file paths

This is to help reduce the duplicating of include files found.  Also fixed opt_v to not require a argument

* New template layout

This new template is to allow testing of new options, and adds coverage option to the build.

* mkmf: fix verbose flag

* fremake & mkmf: only doF90cpp land when source changes

* More templates converted to new style.

* Removed old templates

* Rewrite list_paths to use POSIX shell (sh)

* Updated Make template files

Mostly final version of the make templates.  Some testing still needs to be done to verify these work on the systems indicated, but the layout is now as desired.

* Fix the compiler commands for Fortran, C, and C++

Corrected the make macros used for Fortran and C++ files.

* Correct OPENMP option

Changed ifneq to ifdef for OPENMP macro

* Added FPPFLAGS to .f and .f90 rules

This is needed in the case where `include something` is done to find possible Fortran include files

* Added NetCDF include directories to C and Fortran macros

* More corrections to new templates

Changes required after testing

Updated Cray template to remove openmp if OPENMP is not defined

* More corrections to new templates

Changes required after testing

Updated Cray template to remove openmp if OPENMP is not defined

* Increased version number for next release

* Correct followSymLinks var to indicate when to find symlinked files

* Add basic testing

Added two simple scripts for testing mkmf and list_paths

Configured travis CI automated testing

* Changed build badge to point to NOAA-GFDL/mkmf

* Expanded mkmf and tools tests

Using bats to run tests, with prove to launch tests and interpret the TAP output

* Correct LIBS in gnu-intel template

Missing a `LIBS +=` before the mpich-f90 lookup.

* Correct package name for mpich C

The pkg_config package name for C is `mpich` not `mpich-c`.