Restore commits from find.symlinks branch (#17)
* Updates for POSIX sed

This update allows systems that use a POSIX version of sed (Max OS X).

* list_paths now finds symlinks as well

list_paths used to only find files, this update allows list_paths to find symlinks as well.

* Better fix for symlinks

Better fix for symlinks

This is a better fix for finding symlinks.  This uses the `find` option "-L" to follow the symlink, but return the path using the symlink.  Although, it may be better to use the "-H" option to return the path to the linked file as it will keep `list_path` from returning the two identical files if the symlink points to a file in the same search path.

* Cleanup whitespace/tabs in list_paths

* Remove all ^./ from file paths

This is to help reduce the duplicating of include files found.  Also fixed opt_v to not require a argument

* New template layout

This new template is to allow testing of new options, and adds coverage option to the build.

* mkmf: fix verbose flag

* fremake & mkmf: only doF90cpp land when source changes

* More templates converted to new style.

* Removed old templates

* Rewrite list_paths to use POSIX shell (sh)

* Updated Make template files

Mostly final version of the make templates.  Some testing still needs to be done to verify these work on the systems indicated, but the layout is now as desired.

* Fix the compiler commands for Fortran, C, and C++

Corrected the make macros used for Fortran and C++ files.

* Correct OPENMP option

Changed ifneq to ifdef for OPENMP macro

* Added FPPFLAGS to .f and .f90 rules

This is needed in the case where `include something` is done to find possible Fortran include files

* Added NetCDF include directories to C and Fortran macros

* More corrections to new templates

Changes required after testing

Updated Cray template to remove openmp if OPENMP is not defined

* More corrections to new templates

Changes required after testing

Updated Cray template to remove openmp if OPENMP is not defined

* Increased version number for next release

* Correct followSymLinks var to indicate when to find symlinked files

* Add basic testing

Added two simple scripts for testing mkmf and list_paths

Configured travis CI automated testing

* Changed build badge to point to NOAA-GFDL/mkmf

* Expanded mkmf and tools tests

Using bats to run tests, with prove to launch tests and interpret the TAP output