list_paths to ignore test directories by default
With the addition of automated testing in some of the FMS repositories, it is now important to ignore by default directories that contain test programs.

This update will ignore files in directories that are "t" or "test_.*".  This can be overridden if the "-t" option is given to list_paths.