26 escape.git.version.string (#27)
* Workaround for gfortran not adding -D__APPLE__

Gfortran on Mac OS X does not add the __APPLE__ macro.  Updated the two Mac OS X templates to add this macro definition to the the Fortran preprocessor flags.

* Escape call to git-version-string in mkmf

Update the `-g` option to escape the double quote (`"`) in `mkmf` around 
the call to `git-version-string`.  This is to resolve the quote 
expansion of certain MPI compiler wrappers.

* Update git version test

The test for the git version string needed to be updated to match the update in mkmf to escape the string.