Submodule parsing support; hashing fix
This patch adds support for identifying submodules and tagging the
parent module as their dependency.

Expressions as `submodule (some_mod) some_submod` will tag the object
file of `some_mod` as a dependency of the source file of `some_submod`.

---

This patch also modifies the module regexp to reject any of the
following:

- module procedure
- module subroutine
- module function

The latter two are used when working with submodules.  The first is
typically used for generic interfaces.

Previously the `procedure` exception was applied to the consistency
check, but it may have inadvertently been adding a module named
`procedure` to the hashes.  This was probably not causing any errors,
but was probably nonetheless not the desired behavior.

This change prevents any of the subprogram keywords from being added to
the hashes.