#42 Remove MAKEFLAGS that attempt to set the make --jobs value based on the number of cores available.
This setting is undesired and potentially dangerous for batch jobs,
as it assumes the job can use all cores available. Instead, for batch
jobs the paralellism should be the number of cores allocated to the job.
(Updates to FRE do this.) For interactive jobs, users can set
MAKEFLAGS themselves (e.g. in their shell init files) or as
the make -j option.