Add template files for Gaea C5 cluster
This patch adds new template files for the Gaea C5 cluster.  These are
currently based on the templates used at NOAA-GFDL/mkmf.

Although the PGI compiler has been rechristened as the Nvidia compiler,
we leave the PGI name for now, due to the many hard-coded instances of
the `pgi` label and dependence on those names.  But we may go back and
rename these to nvhpc or something similar.